Code Review Feedback - Part 1
