+++ src/BuiltInTools/Watch/Context/EnvironmentOptions.cs
@@ -40,7 +40,7 @@ namespace Microsoft.DotNet.Watch
         string? AutoReloadWebSocketHostName = null,
         int? AutoReloadWebSocketPort = null,
         string? BrowserPath = null,
-        int HotReloadHttpPort = 0,
+        int AgentWebSocketPort = 0,
         TestFlags TestFlags = TestFlags.None,
         string TestOutput = "")
     {
@@ -60,7 +60,7 @@ namespace Microsoft.DotNet.Watch
             AutoReloadWebSocketHostName: EnvironmentVariables.AutoReloadWSHostName,
             AutoReloadWebSocketPort: EnvironmentVariables.AutoReloadWSPort,
             BrowserPath: EnvironmentVariables.BrowserPath,
-            HotReloadHttpPort: EnvironmentVariables.HotReloadHttpPort,
+            AgentWebSocketPort: EnvironmentVariables.AgentWebSocketPort,
             TestFlags: EnvironmentVariables.TestFlags,
             TestOutput: EnvironmentVariables.TestOutputDir
         );
