+++ src/BuiltInTools/Watch/Context/EnvironmentVariables.cs
@@ -17,7 +17,7 @@ internal static class EnvironmentVariables
 
         public const string DotNetWatchHotReloadNamedPipeName = HotReload.AgentEnvironmentVariables.DotNetWatchHotReloadNamedPipeName;
         public const string DotNetWatchHotReloadWebSocketEndpoint = HotReload.AgentEnvironmentVariables.DotNetWatchHotReloadWebSocketEndpoint;
-        public const string DotNetWatchHotReloadHttpPort = "DOTNET_WATCH_HOTRELOAD_HTTP_PORT";
+        public const string DotNetWatchAgentWebSocketPort = "DOTNET_WATCH_AGENT_WEBSOCKET_PORT";
         public const string DotNetStartupHooks = HotReload.AgentEnvironmentVariables.DotNetStartupHooks;
         public const string DotNetModifiableAssemblies = HotReload.AgentEnvironmentVariables.DotNetModifiableAssemblies;
         public const string HotReloadDeltaClientLogMessages = HotReload.AgentEnvironmentVariables.HotReloadDeltaClientLogMessages;
@@ -63,10 +63,10 @@ internal static class EnvironmentVariables
     public static string? BrowserPath => Environment.GetEnvironmentVariable("DOTNET_WATCH_BROWSER_PATH");
 
     /// <summary>
-    /// Port for HTTP hot reload communication. Used for projects with the HotReloadWebSockets capability.
+    /// Port for WebSocket hot reload communication. Used for projects with the HotReloadWebSockets capability.
     /// Mobile workloads (Android, iOS) add this capability. Defaults to 0 (auto-assign) if not specified.
     /// </summary>
-    public static int HotReloadHttpPort => ReadInt(Names.DotNetWatchHotReloadHttpPort) ?? 0;
+    public static int AgentWebSocketPort => ReadInt(Names.DotNetWatchAgentWebSocketPort) ?? 0;
 
     private static bool ReadBool(string variableName)
         => ParseBool(Environment.GetEnvironmentVariable(variableName));
